Fixed incorrect date format
BadRequestException was being thrown due to "Wrong format of time_start parameter" in Ohlcv_historical_data(string, string, DateTime, DateTime, int). The data format was hard-coded to "yyyy-MM-dd HH:ss:mm", which is invalid. It was changed to dateFormat, which is what it appears to have been changed to in all other methods.